pyproject.toml
[project]
name = "desafio-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "autogen-agentchat==0.2.36",
]
